feat: remove download icon
