+++ src/components/shareDialog/index.js
@@ -9,8 +9,6 @@ import Typography from "@material-ui/core/Typography";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
 import { makeStyles } from "@material-ui/core/styles";
-import SyntaxHighlighter from "react-syntax-highlighter";
-import { a11yDark } from "react-syntax-highlighter/dist/esm/styles/hljs";
 import Button from "@material-ui/core/Button";
 import TextField from "@material-ui/core/TextField";
 import copy from "copy-to-clipboard";
@@ -170,23 +168,6 @@ function CenteredGrid({ params = "" }) {
   );
 }
 
-const MarkdownLink = ({ params = "" }) => {
-  return (
-    <div>
-      <p>
-        You can include the chart on your repository's README.md as follows:
-      </p>
-      <SyntaxHighlighter language="markdown" style={a11yDark}>
-        {`
-## Contributor over time
-
-[![Contributor over time](${IMG_BASE_URL + params})[${SHARE_BASE_URL +
-          params}]]`}
-      </SyntaxHighlighter>
-    </div>
-  );
-};
-
 const DialogTitle = withStyles(styles)(props => {
   const { children, classes, onClose, ...other } = props;
   return (
@@ -232,7 +213,6 @@ export default function CustomizedDialogs({
         <DialogContent dividers>
           <CenteredGrid params={params} />
           <ShareLink params={params} />
-          <MarkdownLink params={params} />
         </DialogContent>
       </Dialog>
     </div>
